fix(engine): isolate performed facts I/O service outside pure engine
